Add support for TravisCI
Adds a .travis.yml file that includes support for 4 of the JDK's available in TravisCI to verify that the examples do not have specific needs.

Needs to be fixed to download the Android SDK and install it before running the relevant tests.

Also needs to be optimised to avoid downloading gradle multiple times.